third_party/blink/web_tests/external/wpt/editing/run/forecolor_1001-2000-expected.txt

This is a testharness.js-based test.
Found 22 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "<span style=\\"color: blue\\">foo<span style=\\"color: brown\\">[bar]</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foo</span><font color=\\"#0000ff\\">bar</font><span style=\\"color:rgb(0, 0, 255)\\">baz</span>" but got "<font color=\\"#0000ff\\">foobarbaz</font>"
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "<span style=\\"color: #00f\\">foo<span style=\\"color: brown\\">[bar]</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foobarbaz</span>" but got "<font color=\\"#0000ff\\">foobarbaz</font>"
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "<span style=\\"color: #0000ff\\">foo<span style=\\"color: brown\\">[bar]</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foobarbaz</span>" but got "<font color=\\"#0000ff\\">foobarbaz</font>"
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "<span style=\\"color: rgb(0, 0, 255)\\">foo<span style=\\"color: brown\\">[bar]</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foobarbaz</span>" but got "<font color=\\"#0000ff\\">foobarbaz</font>"
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "<font color=blue>foo<font color=brown>[bar]</font>baz</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<font color=\\"blue\\">foobarbaz</font>" but got "<font color=\\"#0000ff\\">foo</font><font style=\\"\\" color=\\"#0000ff\\">bar</font><font color=\\"#0000ff\\">baz</font>"
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "<span style=\\"color: rgb(0, 0, 255)\\">foo<span style=\\"color: brown\\">b[ar]</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foo<span style=\\"color:rgb(165, 42, 42)\\">b</span>arbaz</span>" but got "<span style=\\"color:rgb(0, 0, 255)\\">foo</span><span style=\\"color:rgb(165, 42, 42)\\">b</span><span style=\\"color:rgb(0, 0, 255)\\">arbaz</span>"
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "<span style=\\"color: rgb(0, 0, 255)\\">foo<span style=\\"color: brown\\">b[ar]</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foo<font color=\\"#a52a2a\\">b</font>arbaz</span>" but got "<font color=\\"#0000ff\\">foo</font><span style=\\"color:rgb(165, 42, 42)\\">b</span><font color=\\"#0000ff\\">arbaz</font>"
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "foo<span id=purple>ba[r</span>ba]z" queryCommandIndeterm("forecolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "foo<span id=purple>ba[r</span>ba]z" queryCommandIndeterm("forecolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "<span style=\\"color: rgb(0, 0, 255)\\">foo<span id=purple>b[a]r</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foo<span id=\\"purple\\">b<span style=\\"color:rgb(0, 0, 255)\\">a</span>r</span>baz</span>" but got "<span style=\\"color:rgb(0, 0, 255)\\">foo</span><span id=\\"purple\\" style=\\"\\"><span style=\\"color:rgb(0, 0, 255)\\">bar</span></span><span style=\\"color:rgb(0, 0, 255)\\">baz</span>"
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "<span style=\\"color: rgb(0, 0, 255)\\">foo<span id=purple>b[a]r</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foo<span id=\\"purple\\">b<font color=\\"#0000ff\\">a</font>r</span>baz</span>" but got "<font color=\\"#0000ff\\">foo</font><span id=\\"purple\\" style=\\"\\"><font color=\\"#0000ff\\">bar</font></span><font color=\\"#0000ff\\">baz</font>"
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "<font color=\\"0000ff\\">[foo]</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foo</span>" but got "<font color=\\"0000ff\\">foo</font>"
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "<font color=\\"0000ff\\">[foo]</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<font color=\\"#0000ff\\">foo</font>" but got "<font color=\\"0000ff\\">foo</font>"
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "<font color=\\"#0000ff\\">[foo]</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(0, 0, 255)\\">foo</span>" but got "<font color=\\"#0000ff\\">foo</font>"
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "fo[o<font color=brown>b]ar</font>baz" queryCommandIndeterm("forecolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "fo[o<font color=brown>b]ar</font>baz" queryCommandIndeterm("forecolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "foo<font color=brown>ba[r</font>b]az" queryCommandIndeterm("forecolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "foo<font color=brown>ba[r</font>b]az" queryCommandIndeterm("forecolor") before
  assert_equals: Wrong result returned expected true but got false
Harness: the test ran to completion.
